--- examples/gopigo3/mygopigo3/test_tts_wait.urml.yaml
@@ -1,12 +1,0 @@
-profile: educational
-behavior:
-  type: sequence
-  on_error: abort_and_report
-  steps:
-  - speak:
-      utterance: This is a test of the emit_speech primitive with the GoPiGo3 URML
-        runtime
-  - wait:
-      duration: 5
-  - speak:
-      utterance: execution after wait beginning
